Move testfs.MinimalProjectFs -> fixtures.MinimalProjectFs
