+++ internal/pkg/cli/dialog/template_instance_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/keboola/keboola-as-code/internal/pkg/dependencies"
+	"github.com/keboola/keboola-as-code/internal/pkg/fixtures"
 	"github.com/keboola/keboola-as-code/internal/pkg/model"
-	"github.com/keboola/keboola-as-code/internal/pkg/utils/testfs"
 	loadState "github.com/keboola/keboola-as-code/pkg/lib/operation/state/load"
 )
 
@@ -19,7 +19,7 @@ func TestAskTemplateInstance_Interactive(t *testing.T) {
 	// Test dependencies
 	dialog, console := createDialogs(t, true)
 	d := dependencies.NewMockedDeps()
-	projectState, err := d.MockedProject(testfs.MinimalProjectFs(t)).LoadState(loadState.Options{LoadLocalState: true}, d)
+	projectState, err := d.MockedProject(fixtures.MinimalProjectFs(t)).LoadState(loadState.Options{LoadLocalState: true}, d)
 	assert.NoError(t, err)
 	branch, _ := projectState.LocalObjects().Get(model.BranchKey{Id: 123})
 
@@ -65,7 +65,7 @@ func TestAskTemplateInstance_Noninteractive_InvalidInstance(t *testing.T) {
 	// Test dependencies
 	dialog, _ := createDialogs(t, true)
 	d := dependencies.NewMockedDeps()
-	projectState, err := d.MockedProject(testfs.MinimalProjectFs(t)).LoadState(loadState.Options{LoadLocalState: true}, d)
+	projectState, err := d.MockedProject(fixtures.MinimalProjectFs(t)).LoadState(loadState.Options{LoadLocalState: true}, d)
 	assert.NoError(t, err)
 	branch, _ := projectState.LocalObjects().Get(model.BranchKey{Id: 123})
 
@@ -91,7 +91,7 @@ func TestAskTemplateInstance_Noninteractive(t *testing.T) {
 	// Test dependencies
 	dialog, _ := createDialogs(t, true)
 	d := dependencies.NewMockedDeps()
-	projectState, err := d.MockedProject(testfs.MinimalProjectFs(t)).LoadState(loadState.Options{LoadLocalState: true}, d)
+	projectState, err := d.MockedProject(fixtures.MinimalProjectFs(t)).LoadState(loadState.Options{LoadLocalState: true}, d)
 	assert.NoError(t, err)
 	branch, _ := projectState.LocalObjects().Get(model.BranchKey{Id: 123})
